D6 TABAN urunun canli zincirinden olculdu: robot 0.2815 / tespit 0.4215
468 parca, %95 GA [0.2399, 0.3285], TP 567 FP 789 FN 2105.
Marka: NIT 0.0032 (1222 GT'den 2 dogru) | SE 0.0000 | UTL 0.1436 | MOR 0.3041 |
SUPU 0.4149 | UPUN 0.5135 | ONV 0.6232 | S+S 0.3934. Makro 0.2995.

D6 ve D7 dagitilan urun icin BENZER zorlukta (0.2815 vs 0.2980) -- yani D6'da
olculen bagil kazanc D7 icin makul bir on gostergedir.

NIT cokusu dagitilan urunde de VAR; benim eklemedigim, sistemin sureklilik
arz eden kor noktasi.

Ayrica sonda_d6_urun.py'ye payli kosu eklendi.

+++ kos_p6_kademe2.py
@@ -223,14 +223,17 @@ def egit(tr, kol, kafes_bloklar=None, s1ler=None):
             k = kisa(s1ler[i])
             if not len(k):
                 continue
-            Ms.append(kafes_matris(d, kafes_bloklar[i], s1ler[i], k))
+            # float32'ye PARCA BASINA cevir: 2583 parca x ~2000 secenek x 95
+            # sutun float64 birikince vstack tepe bellegi 10 GB'a cikiyor.
+            Ms.append(kafes_matris(d, kafes_bloklar[i], s1ler[i],
+                                   k).astype(np.float32))
             Ys.append(d["y"][k])
         else:
-            Ms.append(oz(d, kol))
+            Ms.append(oz(d, kol).astype(np.float32))
             Ys.append(d["y"][taban_satir(d)] if kol == "TABAN" else d["y"])
     if not Ms:
         return None
-    M = np.vstack(Ms).astype(np.float32)
+    M = np.vstack(Ms)
     Y = np.concatenate(Ys)
     M, Y = alt_ornekle(M, Y)
     return yap().fit(M, Y)
